FastSAM-main/api/requirements.txt
# Base-----------------------------------
matplotlib>=3.2.2
opencv-python>=4.6.0
Pillow>=7.1.2
PyYAML>=5.3.1
requests>=2.23.0
scipy>=1.4.1
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.64.0
setuptools
flask
pillow

pandas>=1.1.4
seaborn>=0.11.0

gradio==3.35.2

numpy 
matplotlib
scikit-learn
tensorflow
# Ultralytics-----------------------------------
ultralytics == 8.0.120

git+https://github.com/openai/CLIP.git#egg=clip